Wait 30 seconds for github to create merge commit.
If you call `opp merge` right after someone else updated the branch
you are merging to, `opp` responds with "still being checked by
github" because github needs a few seconds to attempt to merge in the
background, to tell you if the PR is mergeable or not.

This change waits for this to happen for 30s.