Fixes a deadlock in `opp clean`
When there are more than 10 PRs to clean, we will
pull from the results channel only when we have managed
to correctly enqueue all the `cleanPR` goroutines.
But this can happen only if we manage to acquire
enough slots in the semaphore. And the semaphore
releases only when it manages to push a message to
the results channel
=> 💥